@@ -4,17 +4,14 @@ import { bootstrapApplication } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { AppRoutingModule } from './app/app-routing.module';
 import { importProvidersFrom } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
+
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgxPrintModule } from 'ngx-print';
+import { provideHttpClient } from '@angular/common/http';
 
 bootstrapApplication(AppComponent, {
   providers: [
-    importProvidersFrom(
-      AppRoutingModule,
-      HttpClientModule,
-      ReactiveFormsModule,
-      NgxPrintModule 
-    ),
+    provideHttpClient(),
+    importProvidersFrom(AppRoutingModule, ReactiveFormsModule, NgxPrintModule),
   ],
 }).catch((err) => console.error(err));
